Undo pthread_barrierattr_t change.

+++ nptl/sysdeps/unix/sysv/linux/powerpc/bits/pthreadtypes.h
@@ -171,7 +171,7 @@ typedef volatile int pthread_spinlock_t;
 typedef union
 {
   char __size[__SIZEOF_PTHREAD_BARRIER_T];
-  int __align;
+  long int __align;
 } pthread_barrier_t;
 
 typedef union
